\name{conceptmap}
\alias{conceptmap}
\title{Constructing a conceptmap object}
\usage{
conceptmap(x, ...)
}
\arguments{
  \item{x}{-}

  \item{...}{-}
}
\value{
A conceptmap object.
}
\description{
\code{conceptmap} creates an object that encompasses a
concept maps. For actual implementations, see
\code{\link{conceptmap.default}},
\code{\link{conceptmap.igraph}}, or
\code{\link{conceptmap.matrix}}.
}

