\name{get.unified.concepts}
\alias{get.unified.concepts}
\title{Finding all concepts used.}
\usage{
get.unified.concepts(maps)
}
\arguments{
  \item{maps}{A conceptmaps object.}
}
\value{
A vector of strings of each concepts that appears in at
least one of the maps of \code{maps}.
}
\description{
\code{get.unified.concepts} identifies the common superset
of concepts that is used by the maps of a conceptmaps
object.
}

