\name{modify.concepts}
\alias{modify.concepts}
\title{Modify the concepts of concept maps}
\usage{
modify.concepts(x, concept.list, ...)
}
\arguments{
  \item{x}{A conceptmap or conceptmaps object.}

  \item{concept.list}{A list of concepts.}

  \item{...}{-}
}
\value{
A conceptmaps or conceptmap object.
}
\description{
\code{modify.concepts} modifies the list of concept of a
concept map or of all maps of a set. For actual
implementations see
\code{\link{modify.concepts.conceptmap}}, or
\code{\link{modify.concepts.conceptmaps}}.
}

