\name{summary.conceptmaps}
\alias{summary.conceptmaps}
\title{Return basic information of a conceptmaps object}
\usage{
\method{summary}{conceptmaps}(object, ...)
}
\arguments{
  \item{object}{A conceptmaps object.}

  \item{...}{-}
}
\value{
A matrix with one column for each concept map in the set
and the number of concepts, edges, and components of this
map respectively in 3 rows.
}
\description{
\code{summary} returns basic information about a
conceptmaps object
}

