\name{conceptmaps}
\alias{conceptmaps}
\title{Constructing a conceptmaps object}
\usage{
conceptmaps(x, ...)
}
\arguments{
  \item{x}{-}

  \item{...}{-}
}
\value{
A conceptmaps object.
}
\description{
\code{conceptmaps} creates an object that encompasses a set
of concept maps. For actual implementations, see
\code{\link{conceptmaps.default}},
\code{\link{conceptmaps.list}},
\code{\link{conceptmaps.matrix}}.
}

