\name{print.conceptmap}
\alias{print.conceptmap}
\title{Display basic information of a conceptmap object}
\usage{
\method{print}{conceptmap}(x, ...)
}
\arguments{
  \item{x}{A conceptmap object.}

  \item{...}{-}
}
\value{
-
}
\description{
\code{print} displays basic information. For plotting, see
\code{\link{plot.conceptmap}}
}

