\name{print.conceptmaps}
\alias{print.conceptmaps}
\title{Display basic information of a conceptmaps object}
\usage{
\method{print}{conceptmaps}(x, ...)
}
\arguments{
  \item{x}{A conceptmaps object.}

  \item{...}{-}
}
\value{
-
}
\description{
\code{print} displays basic information. For plotting, see
\code{\link{plot.conceptmaps}}
}

