% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_map.r
\name{summary.conceptmap}
\alias{summary.conceptmap}
\title{Return basic information of a conceptmap object}
\usage{
\method{summary}{conceptmap}(object, ...)
}
\arguments{
\item{object}{A conceptmap object.}

\item{...}{-}
}
\value{
A list with the number of concepts, edges and components of the concept map.
}
\description{
\code{summary} returns basic information about a conceptmap object
}
